Modify input function to reset retries properly
As numIncorrect was defined in the general scope, number of
re-tries from previous attempts persisted and the forfeit message was
seen immediately after an incorrect input. Found that having the
variable still defined at the top (check this) but resetting it to 0
after the forfeit alert was the easiest working solution.

Also reduced number of retries to 2